feat: Move `/set-password` to a better place
