+++ packages/backend/src/app.ts
@@ -24,6 +24,7 @@ import accountsRoutes from './routes/accounts.route';
  * */
 import authRoutes from './routes/auth.route';
 import bankDataProvidersRoutes from './routes/bank-data-providers.route';
+import betterAuthExtensionsRoutes from './routes/better-auth-extensions.route';
 import budgetsRoutes from './routes/budgets.route';
 import categoriesRoutes from './routes/categories.route';
 import binanceRoutes from './routes/crypto/binance.route';
@@ -122,13 +123,16 @@ registerAiCategorizationListeners();
  *  Routes include
  * */
 
-// Custom auth routes (set-password, legacy login/register)
-// Must be mounted BEFORE better-auth to take precedence
-app.use(`${API_PREFIX}/auth`, authRoutes);
+// Better-auth extensions (wraps server-side only APIs like setPassword)
+// Must be mounted BEFORE better-auth handler to take precedence
+app.use(`${API_PREFIX}/auth`, betterAuthExtensionsRoutes);
 
-// Mount better-auth handler for all other auth routes
+// Mount better-auth handler for all auth routes
 // This handles: signup, signin, signout, session, oauth callbacks, passkey, etc.
 app.all(`${API_PREFIX}/auth/*`, toNodeHandler(auth));
+
+// Legacy auth routes (will be removed after migration is complete)
+app.use(`${API_PREFIX}/auth`, authRoutes);
 app.use(`${API_PREFIX}/user`, userRoutes);
 app.use(`${API_PREFIX}/users`, usersRoutes);
 app.use(`${API_PREFIX}/accounts`, accountsRoutes);
