+++ packages/backend/src/controllers/auth.controller.ts
@@ -1,4 +1,3 @@
-import { auth } from '@config/auth';
 import { createController } from '@controllers/helpers/controller-factory';
 import * as authService from '@services/auth.service';
 import { z } from 'zod';
@@ -33,23 +32,3 @@ export const register = createController(
 
 // Noop, token validation happens in middlewares
 export const validateToken = createController(z.object({}), async () => {});
-
-/**
- * Set password for OAuth-only users.
- * Wraps the server-side only auth.api.setPassword() method.
- */
-export const setPassword = createController(
-  z.object({
-    body: z.object({
-      newPassword: z.string().min(8),
-    }),
-  }),
-  async ({ body, req }) => {
-    const result = await auth.api.setPassword({
-      body,
-      headers: req.headers as unknown as Headers,
-    });
-
-    return { data: result };
-  },
-);
